% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{assess_data_quality}
\alias{assess_data_quality}
\title{Assess data quality}
\usage{
assess_data_quality(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
List with quality metrics
}
\description{
Assess data quality
}
