% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{auto_generate_schema}
\alias{auto_generate_schema}
\title{Auto-generate schema from data}
\usage{
auto_generate_schema(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
List with schema
}
\description{
Auto-generate schema from data
}
