% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{clear_readflex_cache}
\alias{clear_readflex_cache}
\title{Clear performance cache and stats}
\usage{
clear_readflex_cache(clear_cache = TRUE, clear_stats = TRUE)
}
\arguments{
\item{clear_cache}{Logical whether to clear encoding cache}

\item{clear_stats}{Logical whether to clear performance stats}
}
\description{
Clear performance cache and stats
}
