% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{detect_all_possible_encodings}
\alias{detect_all_possible_encodings}
\title{Detect all possible encodings with confidence scores}
\usage{
detect_all_possible_encodings(file, sample_content)
}
\arguments{
\item{file}{Character path to file}

\item{sample_content}{Character vector of sample lines}
}
\value{
Data frame with encoding and confidence
}
\description{
Detect all possible encodings with confidence scores
}
