% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{detect_column_widths}
\alias{detect_column_widths}
\title{Detect column widths in fixed-width file}
\usage{
detect_column_widths(file, sample_lines = 20)
}
\arguments{
\item{file}{Character path to file}

\item{sample_lines}{Integer number of lines to analyze}
}
\value{
Numeric vector of column widths
}
\description{
Detect column widths in fixed-width file
}
