% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{detect_file_format}
\alias{detect_file_format}
\title{Detect file format from extension and content}
\usage{
detect_file_format(file)
}
\arguments{
\item{file}{Character path to file}
}
\value{
Character format identifier
}
\description{
Detect file format from extension and content
}
