% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_readflex_config}
\alias{get_readflex_config}
\title{Get current readflex configuration}
\usage{
get_readflex_config(key = NULL)
}
\arguments{
\item{key}{Character specific configuration key to retrieve}
}
\value{
List of configuration or specific value
}
\description{
Get current readflex configuration
}
