% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readflex_enhanced.R
\name{get_readflex_info}
\alias{get_readflex_info}
\title{Get processing information from readflex result}
\usage{
get_readflex_info(data)
}
\arguments{
\item{data}{Data frame returned by readflex_enhanced}
}
\value{
List with processing information
}
\description{
Get processing information from readflex result
}
