% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{is_fixed_width_file}
\alias{is_fixed_width_file}
\title{Detect if file is fixed-width format}
\usage{
is_fixed_width_file(file)
}
\arguments{
\item{file}{Character path to file}
}
\value{
Logical whether file appears to be fixed-width
}
\description{
Detect if file is fixed-width format
}
