% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_readflex_config}
\alias{load_readflex_config}
\title{Load configuration from file}
\usage{
load_readflex_config(config_file = NULL, apply_immediately = TRUE)
}
\arguments{
\item{config_file}{Character path to configuration file}

\item{apply_immediately}{Logical whether to apply loaded config}
}
\value{
List with loaded configuration
}
\description{
Load configuration from file
}
