% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{manage_readflex_cache}
\alias{manage_readflex_cache}
\title{Enhanced cache management}
\usage{
manage_readflex_cache(action = "info")
}
\arguments{
\item{action}{Character action to perform ("info", "clear", "optimize")}
}
\value{
Depends on action
}
\description{
Enhanced cache management
}
