% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{print.readflex_benchmark}
\alias{print.readflex_benchmark}
\title{Print method for benchmark results}
\usage{
\method{print}{readflex_benchmark}(x, ...)
}
\arguments{
\item{x}{readflex_benchmark object}

\item{...}{Additional arguments}
}
\description{
Print method for benchmark results
}
