% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.readflex_diagnostic}
\alias{print.readflex_diagnostic}
\title{Print method for diagnostic results}
\usage{
\method{print}{readflex_diagnostic}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{readflex_diagnostic object}

\item{verbose}{Logical whether to show detailed output}

\item{...}{Additional arguments}
}
\description{
Print method for diagnostic results
}
