% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readflex_enhanced.R
\name{print.readflex_info}
\alias{print.readflex_info}
\title{Print method for readflex info}
\usage{
\method{print}{readflex_info}(x, ...)
}
\arguments{
\item{x}{readflex_info object}

\item{...}{Additional arguments}
}
\description{
Print method for readflex info
}
