% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{print.readflex_report}
\alias{print.readflex_report}
\title{Print method for readflex report}
\usage{
\method{print}{readflex_report}(x, ...)
}
\arguments{
\item{x}{readflex_report object}

\item{...}{Additional arguments}
}
\description{
Print method for readflex report
}
