% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{print.readflex_stats}
\alias{print.readflex_stats}
\title{Print method for readflex statistics}
\usage{
\method{print}{readflex_stats}(x, ...)
}
\arguments{
\item{x}{readflex_stats object}

\item{...}{Additional arguments}
}
\description{
Print method for readflex statistics
}
