% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{print.readflex_validation}
\alias{print.readflex_validation}
\title{Print method for validation results}
\usage{
\method{print}{readflex_validation}(x, ...)
}
\arguments{
\item{x}{readflex_validation object}

\item{...}{Additional arguments}
}
\description{
Print method for validation results
}
