% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print_readflex_config}
\alias{print_readflex_config}
\title{Print method for configuration}
\usage{
print_readflex_config(config = NULL)
}
\arguments{
\item{config}{List with readflex configuration}
}
\description{
Print method for configuration
}
