% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{readflex_aggressive_clean}
\alias{readflex_aggressive_clean}
\title{Aggressive cleaning for problematic files}
\usage{
readflex_aggressive_clean(file, ...)
}
\arguments{
\item{file}{Character path to file}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Aggressive cleaning for problematic files
}
