% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_compressed}
\alias{readflex_compressed}
\title{Compressed file reader}
\usage{
readflex_compressed(file, ...)
}
\arguments{
\item{file}{Character path to compressed file}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Compressed file reader
}
