% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{readflex_diagnostic}
\alias{readflex_diagnostic}
\title{Comprehensive file diagnostics}
\usage{
readflex_diagnostic(file, sample_lines = 100)
}
\arguments{
\item{file}{Character path to file}

\item{sample_lines}{Integer number of lines to sample}
}
\value{
List with diagnostic information
}
\description{
Comprehensive file diagnostics
}
