% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_excel}
\alias{readflex_excel}
\title{Excel file reader with encoding detection}
\usage{
readflex_excel(file, sheet = 1, ...)
}
\arguments{
\item{file}{Character path to Excel file}

\item{sheet}{Character or numeric sheet identifier}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Excel file reader with encoding detection
}
