% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_fixed_width}
\alias{readflex_fixed_width}
\title{Fixed-width file reader}
\usage{
readflex_fixed_width(file, widths = NULL, col_names = NULL, ...)
}
\arguments{
\item{file}{Character path to fixed-width file}

\item{widths}{Numeric vector of column widths}

\item{col_names}{Character vector of column names}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Fixed-width file reader
}
