% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_ods}
\alias{readflex_ods}
\title{ODS (OpenDocument Spreadsheet) reader}
\usage{
readflex_ods(file, sheet = 1, ...)
}
\arguments{
\item{file}{Character path to ODS file}

\item{sheet}{Character or numeric sheet identifier}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
ODS (OpenDocument Spreadsheet) reader
}
