% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_psv}
\alias{readflex_psv}
\title{PSV (Pipe-Separated Values) reader}
\usage{
readflex_psv(file, ...)
}
\arguments{
\item{file}{Character path to PSV file}

\item{...}{Additional arguments passed to readflex}
}
\value{
Data frame
}
\description{
PSV (Pipe-Separated Values) reader
}
