% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{readflex_remove_bom}
\alias{readflex_remove_bom}
\title{Remove BOM from file}
\usage{
readflex_remove_bom(file, encoding = "UTF-8", ...)
}
\arguments{
\item{file}{Character path to file}

\item{encoding}{Character encoding to use}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Remove BOM from file
}
