% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{readflex_shiny_server}
\alias{readflex_shiny_server}
\title{Shiny module server for file processing}
\usage{
readflex_shiny_server(id, max_file_size = 100)
}
\arguments{
\item{id}{Character module ID}

\item{max_file_size}{Numeric maximum file size in MB}
}
\value{
Reactive data frame
}
\description{
Shiny module server for file processing
}
