% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{readflex_tsv}
\alias{readflex_tsv}
\title{TSV (Tab-Separated Values) reader}
\usage{
readflex_tsv(file, ...)
}
\arguments{
\item{file}{Character path to TSV file}

\item{...}{Additional arguments passed to readflex}
}
\value{
Data frame
}
\description{
TSV (Tab-Separated Values) reader
}
