% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{reset_readflex_config}
\alias{reset_readflex_config}
\title{Reset readflex configuration to defaults}
\usage{
reset_readflex_config()
}
\value{
Invisible NULL
}
\description{
Reset readflex configuration to defaults
}
