% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_readflex_config}
\alias{save_readflex_config}
\title{Save current configuration to file}
\usage{
save_readflex_config(config_file = NULL)
}
\arguments{
\item{config_file}{Character path to configuration file}
}
\value{
Logical success status
}
\description{
Save current configuration to file
}
