% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{show_readflex_profiles}
\alias{show_readflex_profiles}
\title{Show available regional profiles}
\usage{
show_readflex_profiles()
}
\value{
Invisible data frame with profile information
}
\description{
Show available regional profiles
}
