% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{type_matches}
\alias{type_matches}
\title{Check if types match for validation}
\usage{
type_matches(actual, expected)
}
\arguments{
\item{actual}{Character actual type}

\item{expected}{Character expected type}
}
\value{
Logical whether types match
}
\description{
Check if types match for validation
}
